# jobbing/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Aprende tu mismo API
  description: Application created to provide training courses
  termsOfService: https://www.youtube.com/channel/UCz3p_Kz87WePeRMy6x5CX5w
  contact:
    name: Jalisco con empleo
    url: https://bdalimentos.org/
    email: comunicacionbamx@bdalimentos.org
  license:
    name: Apache 2.0
    url: http://apache.com
  version: 1.0.0
servers:
- url: https://www.jaliscoconempleo.org
  description: Production Server
- url: https://127.0.0.1:4010
  description: Development server (uses test data)
paths:
...
  /avatar/{uid}:
    put:
      tags:
      - User Profiles
      summary: Upload an avatar
      operationId: avatar_put
      parameters:
      - name: uid
        in: path
        description: Unique identifier
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          image/png:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: successfull response
        "400":
          description: BAD_REQUEST
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"BAD_REQUEST\"}"
                x-content-type: application/json
        "401":
          description: UNAUTHORIZED
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"Unauthorized\"}"
                x-content-type: application/json
        "404":
          description: User not found
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"Error 404, page not found.\"}"
                x-content-type: application/json
        "500":
          description: INTERNAL_SERVER_ERROR
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"unexpected error\"}"
                x-content-type: application/json
      x-openapi-router-controller: jobbing.controllers.user_profiles_controller
...
  /userProfile:
    post:
      tags:
      - User Profiles
      description: Creates a user profile
      operationId: save_user_profile
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfile'
        required: true
      responses:
        "201":
          description: successful response
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"successfully created\"}"
                x-content-type: application/json
      x-openapi-router-controller: jobbing.controllers.user_profiles_controller
    put:
      tags:
      - User Profiles
      description: Updates a User attributes
      operationId: update_user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfile'
        required: true
      responses:
        "200":
          description: successful request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        "204":
          description: successful request
        "400":
          description: BAD_REQUEST
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"BAD_REQUEST\"}"
                x-content-type: application/json
        "401":
          description: UNAUTHORIZED
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"Unauthorized\"}"
                x-content-type: application/json
        "404":
          description: User not found
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"Error 404, page not found.\"}"
                x-content-type: application/json
        "500":
          description: INTERNAL_SERVER_ERROR
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"unexpected error\"}"
                x-content-type: application/json
      x-openapi-router-controller: jobbing.controllers.user_profiles_controller
  /userProfile/{uid}:
    get:
      tags:
      - User Profiles
      description: Displays a User defined by ID
      operationId: get_user_profile_by_id
      parameters:
      - name: uid
        in: path
        description: Unique identifier
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: successful request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        "400":
          description: BAD_REQUEST
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"BAD_REQUEST\"}"
                x-content-type: application/json
        "401":
          description: UNAUTHORIZED
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"Unauthorized\"}"
                x-content-type: application/json
        "404":
          description: User not found
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"Error 404, page not found.\"}"
                x-content-type: application/json
        "500":
          description: INTERNAL_SERVER_ERROR
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"unexpected error\"}"
                x-content-type: application/json
      x-openapi-router-controller: jobbing.controllers.user_profiles_controller
  /address:
    post:
      tags:
      - User Profiles
      description: Creates an address
      operationId: save_address_profile
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Address'
        required: true
      responses:
        "201":
          description: successful response
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"successfully created\"}"
                x-content-type: application/json
      x-openapi-router-controller: jobbing.controllers.user_profiles_controller
    put:
      tags:
      - User Profiles
      description: Updates a User attributes
      operationId: update_addres
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Address'
      responses:
        "200":
          description: successful request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
        "204":
          description: successful request
        "400":
          description: BAD_REQUEST
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"BAD_REQUEST\"}"
                x-content-type: application/json
        "401":
          description: UNAUTHORIZED
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"Unauthorized\"}"
                x-content-type: application/json
        "404":
          description: User not found
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"Error 404, page not found.\"}"
                x-content-type: application/json
        "500":
          description: INTERNAL_SERVER_ERROR
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"unexpected error\"}"
                x-content-type: application/json
      x-openapi-router-controller: jobbing.controllers.user_profiles_controller
  /address/{userId}:
    get:
      tags:
      - User Profiles
      description: Displays an Addres of a user
      operationId: get_addres_by_user_id
      parameters:
      - name: userId
        in: path
        description: Unique identifier
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: successful request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
        "400":
          description: BAD_REQUEST
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"BAD_REQUEST\"}"
                x-content-type: application/json
        "401":
          description: UNAUTHORIZED
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"Unauthorized\"}"
                x-content-type: application/json
        "404":
          description: User not found
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"Error 404, page not found.\"}"
                x-content-type: application/json
        "500":
          description: INTERNAL_SERVER_ERROR
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"unexpected error\"}"
                x-content-type: application/json
      x-openapi-router-controller: jobbing.controllers.user_profiles_controller
...
components:
  schemas:
...
    UserProfile:
      required:
      - firstName
      - firstSurname
      - userprofileId
      type: object
      properties:
        userprofileId:
          type: integer
          description: id for index table
          readOnly: true
          example: 123
        firstName:
          type: string
          description: The identifying name of the contact person.
          example: Saturnino
        secondName:
          type: string
          description: Name of the person
        firstSurname:
          type: string
          description: Name of the person
          example: Herrán
        secondSurname:
          type: string
          description: Name of the person
          example: Guinchard
        birthdate:
          type: string
          description: Name of the person
          format: date
        curp:
          type: string
          description: CURP
          example: HEGS870709HASRNT01
        mobileNumber:
          type: string
          description: Mobile number
          example: "4495484301"
        homeNumber:
          type: string
          description: Home number
          example: "4495484301"
        officeNumber:
          type: string
          description: Office telephone number
          example: "4495484301"
        facebookProfile:
          type: string
          description: facebook profile
          example: saturnino.artist
        linkedinProfile:
          type: string
          description: LinkedIn profile
          example: saturnino.herran
        twitterProfile:
          type: string
          description: Twitter profile
          example: saturnino_h
        idImage:
          type: string
          description: Name of the id file
          example: uuid-image.jpg
        status:
          type: integer
          description: Status of the profile
          readOnly: true
          example: 1
        created:
          type: string
          description: Date of registration
          format: date
        updated:
          type: string
          description: Day of latest update
          format: date
          readOnly: true
        credentialsId:
          type: integer
          description: id to credentials table
          example: 651
        orgId:
          type: integer
          description: Foreing key to Organizations
          example: 23
        address:
          $ref: '#/components/schemas/Address'
      example:
        birthdate: 2000-01-23
        address:
          neighbourhoodId: 5
          zipCode: "12345"
          stateId: 32
          addressLine1: Luis
          addressLine2: Luis
          muncipalityId: 7
          addressId: 35647
        mobileNumber: "4495484301"
        created: 2000-01-23
        officeNumber: "4495484301"
        userprofileId: 123
        idImage: uuid-image.jpg
        credentialsId: 651
        orgId: 23
        facebookProfile: saturnino.artist
        twitterProfile: saturnino_h
        firstName: Saturnino
        linkedinProfile: saturnino.herran
        secondSurname: Guinchard
        homeNumber: "4495484301"
        updated: 2000-01-23
        firstSurname: Herrán
        curp: HEGS870709HASRNT01
        secondName: secondName
        status: 1
    Address:
      required:
      - addressId
      - addressLine1
      - muncipalityId
      - neighbourhoodId
      - stateId
      - zipCode
      type: object
      properties:
        addressId:
          type: integer
          description: Unique id
          readOnly: true
          example: 35647
        addressLine1:
          type: string
          description: "Suite/Apartment, Outer Number, Street"
          example: Luis
        addressLine2:
          type: string
          description: "Suite/Apartment, Outer Number, Street"
          example: Luis
        neighbourhoodId:
          type: integer
          description: Colony or Neighbourhood
          example: 5
        muncipalityId:
          type: integer
          description: Muncipality
          example: 7
        zipCode:
          type: string
          description: Zipcode
          example: "12345"
        stateId:
          type: integer
          description: Foreign key for State
          example: 32
      example:
        neighbourhoodId: 5
        zipCode: "12345"
        stateId: 32
        addressLine1: Luis
        addressLine2: Luis
        muncipalityId: 7
        addressId: 35647
...